+++ tplinkrouterc6u/client/c6u.py
@@ -490,14 +490,15 @@ class TplinkBaseRouter(AbstractRouter, TplinkRequest):
         # Get WAN connected status (for DHCP release/renew)
         if self._wan_ipv4_dynamic and data.get('wan_ipv4_conntype') == 'dhcp':
             try:
-                wan_ipv4_dyn_response = self.request(self._url_wan_ipv4_dynamic + '&operation=read', 'operation=read')
+                wan_ipv4_dyn_response = self.request(
+                    f'{self._url_wan_ipv4_dynamic}&operation=read', 'operation=read')
                 if wan_ipv4_dyn_response:
                     status.ewan_connected = wan_ipv4_dyn_response.get('conn_status') == 'connected'
                 else:
                     self._wan_ipv4_dynamic = False
             except Exception:
                 self._wan_ipv4_dynamic = False
-        
+
         easymesh_device_list = None
         if self._easymesh:
             try:
@@ -805,9 +806,10 @@ class TplinkBaseRouter(AbstractRouter, TplinkRequest):
         self.request(self._url_ipv4_dhcps_write, urlencode(payload))
 
     def set_ewan_connect(self, enable: bool) -> None:
+        """Renew (`True`) or release (`False`) Ethernet WAN DHCP lease. WAN must be DHCP."""
         op = 'renew' if enable else 'release'
-        self.request(self._url_wan_ipv4_dynamic + '&operation=' + op, 'operation=' + op)
-    
+        self.request(f'{self._url_wan_ipv4_dynamic}&operation={op}', f'operation={op}')
+
     @staticmethod
     def _str2bool(v) -> bool | None:
         return str(v).lower() in ("yes", "true", "on") if v is not None else None